standalone_app/requirements.txt
numpy
onnxruntime
opencv-python
pillow
